+++ src/pybliotecario/__init__.py
@@ -1,3 +1,1 @@
 __version__ = "2.0.0"
-
-from pybliotecario.Message import Message
