+++ src/pybliotecario/core_loop.py
@@ -1,6 +1,5 @@
 from datetime import datetime
 import os
-from pybliotecario.Message import Message
 import pybliotecario.on_cmd_message as on_cmd_message
 import logging
 
@@ -89,7 +88,9 @@ def main_loop(tele_api, config=None, clear=False):
             chat_id = message.chat_id
             if message.is_command:
                 # Call the selected command and act on the message
-                response = on_cmd_message.act_on_telegram_command(tele_api, message, config)
+                response = on_cmd_message.act_on_telegram_command(
+                    tele_api, message, config
+                )
                 # If response can be sent to the chat, do so
                 # TODO: make this part of a greater BACKEND class
                 if isinstance(response, str):
@@ -109,7 +110,9 @@ def main_loop(tele_api, config=None, clear=False):
                     tele_api.send_message("¡Archivo recibido y guardado!", chat_id)
                     logger.info("File saved to %s", file_path)
                 else:
-                    tele_api.send_message("There was some problem with this, sorry", chat_id)
+                    tele_api.send_message(
+                        "There was some problem with this, sorry", chat_id
+                    )
                     logger.info(message)
                     logger.warning("There was a problem with this update")
             else:
@@ -127,67 +130,3 @@ def main_loop(tele_api, config=None, clear=False):
                 raise e
 
     tele_api.act_on_updates(act_on_message, not_empty=True)
-
-
-#
-# def main_loop(tele_api, config=None, clear=False):
-#     """
-#     This function activates a "listener" and waits for updates from Telegram
-#     No matter what the update is about, we first store the content and then
-#     if it is a command, we act on the command
-#     """
-#     if config:
-#         accepted_user = config["DEFAULT"]["chat_id"]
-#     else:
-#         accepted_user = None
-#     main_folder = config["DEFAULT"]["main_folder"]
-#     # Get updates from Telegram
-#     raw_updates = tele_api.get_updates(not_empty=True)
-#     updates = [Message(update) for update in raw_updates]
-#     # Act on those updates
-#     for update in updates:
-#         logger.info(update.json)
-#         try:
-#             if update.ignore:
-#                 continue
-#             chat_id = update.chat_id
-#             if update.is_command:
-#                 # Calls select command and act on the message
-#                 # the function will receive the whole telegram API so it is allowed to send msgs directly
-#                 # it can choose to send back a response instead
-#                 response = on_cmd_message.act_on_telegram_command(tele_api, update, config)
-#                 # if response is text, or file, it will be sent to the chat
-#                 if isinstance(response, str):
-#                     tele_api.send_message(response, chat_id)
-#                 elif isinstance(response, dict):
-#                     if response["isfile"]:
-#                         filepath = response["filename"]
-#                         tele_api.send_file(filepath, chat_id)
-#                         if response["delete"]:
-#                             os.remove(filepath)
-#             elif update.isFile:
-#                 # If the update is a file, save the file and we are done
-#                 file_name = update.text.replace(" ", "")
-#                 file_path = "{0}/{1}".format(monthly_folder(main_folder), file_name)
-#                 result = tele_api.download_file(update.fileId, file_path)
-#                 if result:
-#                     tele_api.send_message("¡Archivo recibido y guardado!", chat_id)
-#                     logger.info("File saved to {0}".format(file_path))
-#                 else:
-#                     tele_api.send_message("There was some problem with this, sorry", chat_id)
-#                     logger.info(
-#                         "Since there was some problem, let's open a pdb console here and you decide what to do"
-#                     )
-#
-#             else:
-#                 # Otherwise just save the msg to the log and send a funny reply
-#                 write_to_daily_log(main_folder, update.text)
-#                 random_msg = still_alive()
-#                 tele_api.send_message(random_msg, chat_id)
-#         except Exception as e:
-#             # If we are in clear mode, we want to recapture updates to ensure we clear the ones that produce a fail
-#             # in principle in clear mode we don't care what the fail is about, we just want to clear the failure
-#             if clear:
-#                 logger.info("\n     > > This update produced an exception: {0}\n\n".format(e))
-#             else:
-#                 raise e
